Add support for BLE extended advertisements (#263)
* Add support for BLE extended advertisements

* Save information about used BLE PHYs for scanning, enable extended advertisements when 2M PHY or Coded PHY is used

* Collect extra information when receiving advertisement: used primary and secondary PHY, radio channel and tx_power

* Discard regular advertisements from 1M PHY if only 2M PHY is enabled

* Makefile: remove PVS-studio checking

* Doxyfile: remove deprecated options

* Update unit-tests

* Fix doxygen generation

* Makefile: adjust formatting rules to prevent formatting of submodule src/nrf5_sdk15_platform/ruuvi.nrf_sdk15_3_overrides.c

* Fix code smell found by SonarCloud

+++ src/interfaces/communication/ruuvi_interface_communication_radio.h
@@ -49,9 +49,9 @@ typedef enum
  */
 typedef struct
 {
-    unsigned int channel_37 : 1; //!< BLE channel 37, 2402 MHz.
-    unsigned int channel_38 : 1; //!< BLE channel 38, 2426 MHz.
-    unsigned int channel_39 : 1; //!< BLE channel 39, 2480 MHz.
+    uint8_t channel_37 : 1; //!< BLE channel 37, 2402 MHz.
+    uint8_t channel_38 : 1; //!< BLE channel 38, 2426 MHz.
+    uint8_t channel_39 : 1; //!< BLE channel 39, 2480 MHz.
 } ri_radio_channels_t;
 
 /**
